+++ HISTORY.md
@@ -7,6 +7,8 @@
 - Fix issues with concatenating files that fail to work with GATK's
   CatVariants. Fall back to bcftools concat which correctly handles problem
   headers and overlapping segments.
+- Move to samtools 1.0 release. Update samtools variant calling to support new
+  multiallelic approach.
 - Improve Platypus integration: correctly pass multiple BAM files, make use of
   assembler, split MNPs, and correctly restrict to variant regions.
 - Be more aggressive with system memory usage to try and make better use of